+++ src/main/java/de/hysky/skyblocker/mixins/ClientPlayNetworkHandlerMixin.java
@@ -15,7 +15,7 @@ import de.hysky.skyblocker.skyblock.dungeon.secrets.DungeonManager;
 import de.hysky.skyblocker.skyblock.end.BeaconHighlighter;
 import de.hysky.skyblocker.skyblock.end.EnderNodes;
 import de.hysky.skyblocker.skyblock.end.TheEnd;
-import de.hysky.skyblocker.skyblock.slayers.SlayerMobs;
+import de.hysky.skyblocker.skyblock.slayers.SlayerEntitiesGlow;
 import de.hysky.skyblocker.skyblock.waypoint.MythologicalRitual;
 import de.hysky.skyblocker.utils.SlayerUtils;
 import de.hysky.skyblocker.utils.Utils;
@@ -111,7 +111,7 @@ public abstract class ClientPlayNetworkHandlerMixin {
         if (packet.getStatus() == EntityStatuses.PLAY_DEATH_SOUND_OR_ADD_PROJECTILE_HIT_PARTICLES) {
             DungeonScore.handleEntityDeath(entity);
             TheEnd.onEntityDeath(entity);
-            SlayerMobs.onEntityDeath(entity);
+            SlayerEntitiesGlow.onEntityDeath(entity);
         }
         return entity;
     }
@@ -121,9 +121,9 @@ public abstract class ClientPlayNetworkHandlerMixin {
         if (!(entity instanceof ArmorStandEntity armorStandEntity)) return;
 
         // Don't want to waste resources, so if the entities been in the world a while we ignore it. Slightly more lenience on Boss' age because it's more important we don't miss it
-        if (SkyblockerConfigManager.get().slayers.highlightMinis == SlayersConfig.HighlightSlayerEntities.GLOW  && entity.age < 30 && SlayerMobs.isSlayerMiniMob(armorStandEntity)
-        || SkyblockerConfigManager.get().slayers.highlightBosses == SlayersConfig.HighlightSlayerEntities.GLOW && entity.age < 40 && SlayerMobs.isSlayer(armorStandEntity)) {
-            SlayerMobs.setSlayerMobGlow(armorStandEntity);
+        if (SkyblockerConfigManager.get().slayers.highlightMinis == SlayersConfig.HighlightSlayerEntities.GLOW  && entity.age < 30 && SlayerEntitiesGlow.isSlayerMiniMob(armorStandEntity)
+        || SkyblockerConfigManager.get().slayers.highlightBosses == SlayersConfig.HighlightSlayerEntities.GLOW && entity.age < 40 && SlayerEntitiesGlow.isSlayer(armorStandEntity)) {
+            SlayerEntitiesGlow.setSlayerMobGlow(armorStandEntity);
         }
 
         if (SkyblockerConfigManager.get().slayers.blazeSlayer.firePillarCountdown != SlayersConfig.BlazeSlayer.FirePillar.OFF) FirePillarAnnouncer.checkFirePillar(entity);
